# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LASU Fleet Manager API",
        "description": "API documentation for the LASU fleet management system.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/manager/vehicles": {
            "get": {
                "tags": [
                    "Manager"
                ],
                "summary": "List all vehicles",
                "operationId": "3a2fbba0368bf8d8f598fafbd5159a70",
                "responses": {
                    "200": {
                        "description": "Vehicle list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "plateNumber": {
                                                "type": "string",
                                                "example": "LASU-123"
                                            },
                                            "ownerId": {
                                                "type": "integer",
                                                "example": 42
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "active"
                                            },
                                            "model": {
                                                "type": "string",
                                                "example": "Toyota Camry"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/manager/vehicles/{vehicleId}": {
            "put": {
                "tags": [
                    "Manager"
                ],
                "summary": "Update a vehicle's status",
                "operationId": "ef0eb783a0f88692d550868e59321b67",
                "parameters": [
                    {
                        "name": "vehicleId",
                        "in": "path",
                        "description": "Vehicle ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ],
                                        "example": "inactive"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated vehicle status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "inactive"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/owners/{ownerId}/condition-updates": {
            "get": {
                "tags": [
                    "Owner"
                ],
                "summary": "Get condition updates for owner's vehicles",
                "operationId": "708d61f36e070a67d12eda71c933a47c",
                "parameters": [
                    {
                        "name": "ownerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of updates"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Owner"
                ],
                "summary": "Create a condition update",
                "operationId": "6865521927c4666e1fc90a8a09b4854e",
                "parameters": [
                    {
                        "name": "ownerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehicleId",
                                    "condition"
                                ],
                                "properties": {
                                    "vehicleId": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "condition": {
                                        "type": "string",
                                        "example": "Good"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "No visible issues"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/owners/{ownerId}/fuel-requests": {
            "get": {
                "tags": [
                    "Owner"
                ],
                "summary": "Get all fuel requests for owner",
                "operationId": "5ad51aaca2336fe25602c7b81b00c91a",
                "parameters": [
                    {
                        "name": "ownerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List fuel requests"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Owner"
                ],
                "summary": "Submit a new fuel request",
                "operationId": "a000c8f1490781584362e504b82bb603",
                "parameters": [
                    {
                        "name": "ownerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehicleId",
                                    "litres",
                                    "reason"
                                ],
                                "properties": {
                                    "vehicleId": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "litres": {
                                        "type": "number",
                                        "example": 30
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "Weekly operation"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Fuel request created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/owners/{ownerId}/inspection-files": {
            "get": {
                "tags": [
                    "Owner"
                ],
                "summary": "List inspection files for owner",
                "operationId": "767dd50228778a26fbcdad6c41d94b31",
                "parameters": [
                    {
                        "name": "ownerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of uploaded files"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Owner"
                ],
                "summary": "Upload inspection file",
                "operationId": "fabbb407d62ae8c0f41abd7ef6c3b9d1",
                "parameters": [
                    {
                        "name": "ownerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "vehicleId",
                                    "file"
                                ],
                                "properties": {
                                    "vehicleId": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "File uploaded"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}